Reduce integration test flakes (#97)
* Reduce flakes for integration tests

Wait for the principal and resource to be available before trying to
delete.
Also extract some functions to the suite

* Update CHANGELOG

* Remove minor duplication

+++ tests/integration/aws/ram_test.go
@@ -5,8 +5,6 @@ import (
 	"time"
 
 	awssdk "github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials/stscreds"
-	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/ram"
 	. "github.com/onsi/ginkgo/v2"
@@ -26,51 +24,18 @@ var _ = Describe("RAM client", func() {
 		share      aws.ResourceShare
 		prefixList *ec2.ManagedPrefixList
 
-		rawEC2Client *ec2.EC2
-		rawRamClient *ram.RAM
-
 		ramClient *aws.RAMClient
 	)
 
-	getSharedResources := func(g Gomega) []*ram.Resource {
-		listResourcesOutput, err := rawRamClient.ListResources(&ram.ListResourcesInput{
-			Principal:     awssdk.String(wcAccount),
-			ResourceArns:  awssdk.StringSlice([]string{*prefixList.PrefixListArn}),
-			ResourceOwner: awssdk.String(aws.ResourceOwnerSelf),
-		})
-		g.Expect(err).NotTo(HaveOccurred())
-		return listResourcesOutput.Resources
-	}
+	waitForResourceShareAvailability := func() {
+		// The resource share can only be deleted when the resource and
+		// principal are in a final state like Associated. If we don't
+		// wait for this state the tests will flake
+		Eventually(getResourceAssociationStatus(name, prefixList)).
+			Should(PointTo(Equal(ram.ResourceShareAssociationStatusAssociated)))
 
-	getResourceAssociationStatus := func(g Gomega) *string {
-		listAssociationsOutput, err := rawRamClient.GetResourceShareAssociations(&ram.GetResourceShareAssociationsInput{
-			AssociationType: awssdk.String(ram.ResourceShareAssociationTypeResource),
-			ResourceArn:     prefixList.PrefixListArn,
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		Expect(listAssociationsOutput.ResourceShareAssociations).To(HaveLen(1))
-		return listAssociationsOutput.ResourceShareAssociations[0].Status
-	}
-
-	getPrincipalAssociationStatus := func(g Gomega) *string {
-		getResourceShareOutput, err := rawRamClient.GetResourceShares(&ram.GetResourceSharesInput{
-			Name:          awssdk.String(name),
-			ResourceOwner: awssdk.String(aws.ResourceOwnerSelf),
-		})
-		Expect(err).NotTo(HaveOccurred())
-		Expect(getResourceShareOutput.ResourceShares).To(HaveLen(1))
-
-		resourceShare := getResourceShareOutput.ResourceShares[0]
-		listAssociationsOutput, err := rawRamClient.GetResourceShareAssociations(&ram.GetResourceShareAssociationsInput{
-			AssociationType:   awssdk.String(ram.ResourceShareAssociationTypePrincipal),
-			Principal:         awssdk.String(wcAccount),
-			ResourceShareArns: []*string{resourceShare.ResourceShareArn},
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		Expect(listAssociationsOutput.ResourceShareAssociations).To(HaveLen(1))
-		return listAssociationsOutput.ResourceShareAssociations[0].Status
+		Eventually(getPrincipalAssociationStatus(name)).
+			Should(PointTo(Equal(ram.ResourceShareAssociationStatusAssociated)))
 	}
 
 	BeforeEach(func() {
@@ -79,22 +44,6 @@ var _ = Describe("RAM client", func() {
 
 		ctx = log.IntoContext(context.Background(), logger)
 		name = tests.GenerateGUID("test")
-
-		session, err := session.NewSession(&awssdk.Config{
-			Region: awssdk.String(awsRegion),
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		rawEC2Client = ec2.New(session,
-			&awssdk.Config{
-				Credentials: stscreds.NewCredentials(session, mcIAMRoleARN),
-			},
-		)
-		rawRamClient = ram.New(session,
-			&awssdk.Config{
-				Credentials: stscreds.NewCredentials(session, mcIAMRoleARN),
-			},
-		)
 
 		prefixList = createManagedPrefixList(rawEC2Client, name)
 
@@ -107,43 +56,49 @@ var _ = Describe("RAM client", func() {
 	})
 
 	AfterEach(func() {
-		_, err := rawEC2Client.DeleteManagedPrefixList(&ec2.DeleteManagedPrefixListInput{PrefixListId: prefixList.PrefixListId})
-		Expect(err).NotTo(HaveOccurred())
-
 		getResourceShareOutput, err := rawRamClient.GetResourceShares(&ram.GetResourceSharesInput{
 			Name:          awssdk.String(name),
 			ResourceOwner: awssdk.String(aws.ResourceOwnerSelf),
 		})
 		Expect(err).NotTo(HaveOccurred())
 		for _, resourceShare := range getResourceShareOutput.ResourceShares {
+			if isResourceShareDeleted(resourceShare) {
+				continue
+			}
+
 			_, err = rawRamClient.DeleteResourceShare(&ram.DeleteResourceShareInput{
 				ResourceShareArn: resourceShare.ResourceShareArn,
 			})
 			Expect(err).NotTo(HaveOccurred())
 		}
+
+		_, err = rawEC2Client.DeleteManagedPrefixList(&ec2.DeleteManagedPrefixListInput{PrefixListId: prefixList.PrefixListId})
+		Expect(err).NotTo(HaveOccurred())
 	})
 
 	Describe("ApplyResourceShare", func() {
 		It("creates the share resource", func() {
 			err := ramClient.ApplyResourceShare(ctx, share)
 			Expect(err).NotTo(HaveOccurred())
+			waitForResourceShareAvailability()
 
-			Eventually(getSharedResources).Should(HaveLen(1))
+			Eventually(getSharedResources(rawRamClient, prefixList)).Should(HaveLen(1))
 		})
 
 		When("the resource has already been shared", func() {
 			BeforeEach(func() {
 				err := ramClient.ApplyResourceShare(ctx, share)
 				Expect(err).NotTo(HaveOccurred())
+				waitForResourceShareAvailability()
 
-				Eventually(getSharedResources).Should(HaveLen(1))
+				Eventually(getSharedResources(rawRamClient, prefixList)).Should(HaveLen(1))
 			})
 
 			It("does not return an error", func() {
 				err := ramClient.ApplyResourceShare(ctx, share)
 				Expect(err).NotTo(HaveOccurred())
 
-				Consistently(getSharedResources).Should(HaveLen(1))
+				Consistently(getSharedResources(rawRamClient, prefixList)).Should(HaveLen(1))
 			})
 		})
 
@@ -189,20 +144,16 @@ var _ = Describe("RAM client", func() {
 		BeforeEach(func() {
 			err := ramClient.ApplyResourceShare(ctx, share)
 			Expect(err).NotTo(HaveOccurred())
-			Eventually(getSharedResources).Should(HaveLen(1))
+			Eventually(getSharedResources(rawRamClient, prefixList)).Should(HaveLen(1))
 
-			// The resource share can only be deleted when the resource and
-			// principal are in a final state like Associated. If we don't
-			// wait for this state the tests will flake
-			Eventually(getResourceAssociationStatus).Should(PointTo(Equal(ram.ResourceShareAssociationStatusAssociated)))
-			Eventually(getPrincipalAssociationStatus).Should(PointTo(Equal(ram.ResourceShareAssociationStatusAssociated)))
+			waitForResourceShareAvailability()
 		})
 
 		It("deletes the resource share", func() {
 			err := ramClient.DeleteResourceShare(ctx, name)
 			Expect(err).NotTo(HaveOccurred())
 
-			Eventually(getSharedResources).Should(HaveLen(0))
+			Eventually(getSharedResources(rawRamClient, prefixList)).Should(HaveLen(0))
 		})
 
 		When("the resource is already deleted", func() {
@@ -210,7 +161,7 @@ var _ = Describe("RAM client", func() {
 				err := ramClient.DeleteResourceShare(ctx, name)
 				Expect(err).NotTo(HaveOccurred())
 
-				Eventually(getSharedResources).Should(HaveLen(0))
+				Eventually(getSharedResources(rawRamClient, prefixList)).Should(HaveLen(0))
 			})
 
 			It("does not return an error", func() {
@@ -220,18 +171,18 @@ var _ = Describe("RAM client", func() {
 
 			When("creating a resource share with the same name", func() {
 				It("creates the share resource", func() {
-					Consistently(getSharedResources).Should(HaveLen(0))
-
 					err := ramClient.ApplyResourceShare(ctx, share)
 					Expect(err).NotTo(HaveOccurred())
 
-					Eventually(getSharedResources).Should(HaveLen(1))
+					Eventually(getSharedResources(rawRamClient, prefixList)).Should(HaveLen(1))
+					waitForResourceShareAvailability()
 				})
 
 				When("the resource has already been recreated", func() {
 					BeforeEach(func() {
 						err := ramClient.ApplyResourceShare(ctx, share)
 						Expect(err).NotTo(HaveOccurred())
+						waitForResourceShareAvailability()
 					})
 
 					It("does not return an error", func() {
